fastapi_service/requirements.txt
uvicorn==0.29.0
fastapi==0.110.1
pymongo==4.6.3
email-validator==2.1.1
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
PyJWT==2.8.0